requirements.txt
aiohttp==3.6.2
aioredis==1.3.1
async-timeout==3.0.1
attrs==19.3.0
bitmex-ws==0.4.0
certifi==2020.4.5.1
chardet==3.0.4
dnspython==1.16.0
hiredis==1.0.1
idna==2.9
motor==2.1.0
multidict==4.7.5
pymongo==3.10.1
redis==3.4.1
six==1.14.0
sortedcontainers==2.1.0
websocket-client==0.53.0
yarl==1.4.2
